用户删除及接口验证补全

+++ routes/role.js
@@ -44,8 +44,12 @@ router.post('/deleteRole', async function (req, res) {
 // 根据用户id获取用户角色
 router.get('/getRoleByUserId', async function (req, res) {
   const {pk_user_id} = req.query
-  const data = await roleService.getRoleByUserId(pk_user_id);
-  res.send({status: 0, data: data});
+  if (!validator.isInt(pk_user_id)) {
+    res.send({status: 1, msg: '用户id格式不正确，请检查后重新输入'})
+  } else {
+    const data = await roleService.getRoleByUserId(pk_user_id);
+    res.send({status: 0, data: data});
+  }
 });
 
 /*
@@ -57,40 +61,50 @@ router.get('/getRoleByUserId', async function (req, res) {
 router.post('/updateRoleRights', async function (req, res) {
   let {pk_role_id, menus} = req.body
   console.log(pk_role_id, menus)
-  const rights = await rightService.getRightsByRoleId(pk_role_id)
+  if (!validator.isInt(pk_user_id)) {
+    res.send({status: 1, msg: '用户id格式不正确，请检查后重新输入'})
+  } else {
+    const rights = await rightService.getRightsByRoleId(pk_role_id)
 
-  let oldMenus = []
-  for (let i = 0; i < rights.length; i++) {
-    oldMenus.push(rights[i].fk_right_id)
+    let oldMenus = []
+    for (let i = 0; i < rights.length; i++) {
+      oldMenus.push(rights[i].fk_right_id)
+    }
+
+    let newMenus = []
+    menus.toString().split(",").forEach(function (item) {
+      newMenus.push(parseInt(item));
+    })
+    console.log(newMenus, oldMenus)
+
+    oldMenus.forEach(async function (item) {
+      if (newMenus.indexOf(item) === -1) {
+        console.log(item)
+        await rightService.deleteRoleRights(pk_role_id, item)
+      }
+    })
+
+    newMenus.forEach(async function (item) {
+      if (oldMenus.indexOf(item) === -1) {
+        console.log(item)
+        await rightService.addRoleRights(pk_role_id, item)
+      }
+    })
+    res.send({status: 0});
   }
-
-  let newMenus = []
-  menus.toString().split(",").forEach(function (item) {
-    newMenus.push(parseInt(item));
-  })
-  console.log(newMenus, oldMenus)
-
-  oldMenus.forEach(async function (item) {
-    if (newMenus.indexOf(item) === -1) {
-      console.log(item)
-      await rightService.deleteRoleRights(pk_role_id, item)
-    }
-  })
-
-  newMenus.forEach(async function (item) {
-    if (oldMenus.indexOf(item) === -1) {
-      console.log(item)
-      await rightService.addRoleRights(pk_role_id, item)
-    }
-  })
-  res.send({status: 0});
 })
 
 // 更新用户的角色
 router.post('/updateRoleByUserId', async function (req, res) {
   const {pk_user_id, role_id} = req.body
-  const data = await roleService.updateRoleByUserId(pk_user_id, role_id);
-  res.send(data);
+  if (!validator.isInt(pk_user_id)) {
+    res.send({status: 1, msg: '用户id格式不正确，请检查后重新输入'})
+  } else if (!validator.isInt(role_id)) {
+    res.send({status: 1, msg: '角色id格式不正确，请检查后重新输入'})
+  } else {
+    const data = await roleService.updateRoleByUserId(pk_user_id, role_id);
+    res.send(data);
+  }
 })
 
 module.exports = router;
